Propose consolidation refactor
