+++ CauldronTests/Parsing/SocialRecipeParserArchitectureTests.swift
@@ -1,5 +1,5 @@
 //
-//  YouTubeRecipeParserTests.swift
+//  SocialRecipeParserArchitectureTests.swift
 //  CauldronTests
 //
 //  Created on November 13, 2025.
@@ -8,15 +8,15 @@
 import XCTest
 @testable import Cauldron
 
-/// Tests for YouTube recipe parser
+/// Tests for social recipe parser architecture
 ///
-/// NOTE: YouTubeRecipeParser is an actor with private methods. Most of the parsing logic
+/// NOTE: SocialRecipeParser is an actor with private methods. Most of the parsing logic
 /// has been extracted into testable utilities (QuantityValueParser, UnitParser,
 /// IngredientParser, TextSectionParser, TimerExtractor) which are thoroughly tested.
 ///
-/// These tests focus on integration and any YouTubeRecipeParser-specific logic.
+/// These tests focus on architecture-level assertions.
 @MainActor
-final class YouTubeRecipeParserTests: XCTestCase {
+final class SocialRecipeParserArchitectureTests: XCTestCase {
 
     // MARK: - Architecture Verification
 
@@ -29,22 +29,20 @@ final class YouTubeRecipeParserTests: XCTestCase {
         XCTAssertNotNil(TimerExtractor.self)
     }
 
-    func testYouTubeParserExists() {
-        // Verify that YouTubeRecipeParser compiles and exists
-        // Note: Since it's an actor, direct instantiation requires FoundationModelsService
-        XCTAssertTrue(true, "YouTubeRecipeParser exists and uses tested utility parsers")
+    func testSocialParserExists() {
+        XCTAssertNotNil(SocialRecipeParser.self)
     }
 
     // MARK: - Integration Notes
 
-    /// The YouTubeRecipeParser integrates the following tested components:
+    /// The SocialRecipeParser integrates the following tested components:
     /// ✅ QuantityValueParser - Tested with 25 tests (all passing)
     /// ✅ UnitParser - Tested with 38 tests (all passing)
     /// ✅ IngredientParser - Tested with 25 tests (all passing)
     /// ✅ TextSectionParser - Tested with 29 tests (all passing)
     /// ✅ TimerExtractor - Tested with 21 tests (all passing)
     ///
-    /// Total: 138 tests covering the core parsing logic used by YouTubeRecipeParser
+    /// Total: 138+ tests covering the core parsing logic used by social import parsing
     ///
     /// Future improvements:
     /// - Add mock URLSession for integration tests with sample YouTube HTML
